Add support for connecting to Calpendo 11.0 and later
From version 11.0 onwards, the cookie used to store the user's session ID
has a different name. This is done as a part of supporting Calpendo
Enterprise, in which a federation of multiple Calpendo shards work
together.

For this to work properly, each Calpendo shard needs a different cookie
for storing its session ID, so the cookie name now includes the shard
number.

We now add a script that fetches the name of the cookie, and then uses it
in the call to run a workflow.

The scripts are backwards-compatible with older Calpendo systems.